Fix magnifier pixel value display
Apparently, when using XShmGetImage to get an XImage for a window,
the image no longer includes subwindows (like when using
IncludeInferiors in GC).
When using XGetImage the XImage still includes subwindows.

Changing EImageGrabDrawable() to EImageGrabDrawableScaled() means that
we use imlib_create_scaled_image_from_drawable() instead of
imlib_create_image_from_drawable().

imlib_create_scaled_image_from_drawable() is implemented in such a way
that the drawable to be grabbed is always copied to a pixmap first.