Handle ".theme files"
A .theme file is really the same as the usual .etheme file (i.e.
a tarball of the theme files, compressed or not), except that .theme
files may contain themes for multiple applications.

The difference in the handling of a .theme file is only the location
where it is extracted.

A normal theme file, say FOO.etheme (or FOO.tar, FOO.tar.gz, FOO.blah),
is (normally) extracted to ~/.e16/themes/FOO/.
e16 now expects the theme files to be located in
~/.e16/themes/FOO/*.cfg or ~/.e16/themes/FOO/e16/*.cfg.

A .theme file, e.g. FOO.theme, is extracted to ~/.themes/FOO/.
e16 now expects the theme files to be located in
~/.themes/FOO/*.cfg or ~/.themes/FOO/e16/*.cfg (the first one
should probably never be used, but it's how it works).

In order to find this new theme path it is added to the default theme
search path.

Suggested by Don Harrop.