Backgrounds: Fix bad list operation in CB_ConfigureDelBG()
If ever we would delete the last background in the list we would get a
segv.
However, this will never occur as the None background is always the last
in the list and we will never delete it.

This can be a bit difficult to comprehend for static analysis tools so
eliminate the offending list operation.