edox: Don't try rendering text outside edox window
When a text line to be rendered does not fit at the current position,
the corrected position could be set outside the edox window, which would
cause errors reported by imlib2.

Now the corrected position will be set to the lower left corner. This
may cause text to not be rendered exactly where it should be, but at
least it is rendered (which makes it easier to spot the problem), and
imlib2 errors are avoided.