Disable setting internal window title via ipc (properly)
Previously, when trying to set the title via ipc, the window WM_NAME and
_NET_WM_NAME properties would be changed, but the border title (e16's
perception of the properties) would not be updated because PropertyChange
notifications for internal windows are discarded (to avoid processing
changes we already know have happened, IIRC).

Now the internal window name properties are not changed, avoiding
inconsistency between properties and internal state (and border title).