Clamp window position on move
In non-compositing mode the x/y coordinate limts are [-32768; 32767] as
should be expected. If moving beyond these values the coordinates are
just set modulo 2^16 (like in a signed 16 bit integer) and most likely
this isn't noticed anywhere.

In compositing mode the coordinates apperently must be clamped some
more, but it is not clear why or where the limit is.
The problem manifests itself in that a window moved "too far" seems to
have disappeared when moving it back on screen, possibly because the
name pixmap is invalidated.
Shading/unshading the window makes it re-appear.

The problem can be seen when dragging a window far off a (small) pager.