menus: Fix potential crash
When requesting to show a menu already shown it would be hidden before
being shown again, which could cause a crash due to incorrect processing
of reparent events.

Now we just ignore requests to show a menu that is aleady shown, which
bypasses the event processing issue and is reasonable anyway.