ewins: Tweak spurious MapRequest handling
Some clients produce multiple MapRequests (presumably by doing
multiple XMapWindow()s early, which hits case "B").
Not sure when case "A" is hit, maybe in connection with incorrect
iconification or withdrawal and re-map.

Now we only reparent (and move) if not already reparented, and place
correcly in the no-container case (which most likely is to do nothing
as the window probably already is parented/placed correctly).
Also always print a message to the console on spurious MapRequests.

This fixes client window placement within the frame in the no-container
case with some clients.