iclass: Drop image border adjustment
In the switch from imlib to imlib2 (a couple of decades ago) an
adjustment of image borders was introduced to accommodate for a
difference between the old and new imlibs.

Without this adjustment a few themes have incorrectly rendered elements
(unintended window border fading, opaque pager area highligt frame, more?).

It turns out that the origin of the problem is in the scaling code of
imlib2, and will be fixed in imlib2 (most likely v1.12.0).